pyproject.toml
[tool.poetry]
name = "repl_python3_BotMb-Telegram"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
telebot = "^0.0.4"
pyTelegramBotAPI = "^4.1.1"
redis = "^3.5.3"
ujson = "^4.2.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
